+++ app/src/SubmitterInformation.tsx
@@ -13,7 +13,7 @@ interface IProps {
 export const SubmitterInformation = ({ project_id, personal }: IProps) => {
     const subject = "Regarding paired omics data platform project: " + project_id;
     const submitter_mailto = "mailto:" + personal.submitter_email + "?subject=" + subject;
-    const pi_mailto = "mailto:" + personal.submitter_email + "?subject=" + subject;
+    const pi_mailto = "mailto:" + personal.PI_email + "?subject=" + subject;
     return (
         <ul>
             <li>