+++ app/src/DownloadPendingProject.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { Button, Glyphicon } from 'react-bootstrap';
 import { fetchPendingProject } from './api';
 import { saveAs } from 'file-saver';
+import { EnrichedProjectDocument } from './summarize';
 
 interface IProps {
     project_id: string;
@@ -13,7 +14,8 @@ export const DownloadPendingProject = ({project_id, token}: IProps) => {
     const onClick = async () => {
         const response = await fetchPendingProject(project_id, token);
         if (response.ok) {
-            const blob = await response.blob();
+            const json: EnrichedProjectDocument = await response.json();
+            const blob = JSON.stringify(json.project, null, 4);
             const file = new File([blob], filename, { type: 'application/json' });
             saveAs(file);
         } else {