+++ api/src/app.ts
@@ -35,6 +35,7 @@ export function builder(mystore: ProjectDocumentStore, myenrichqueue: Bull.Queue
     app.post('/api/projects/:id', asyncHandler(controller.editProject));
     app.get('/api/projects/:id/history', asyncHandler(controller.getProjectHistory));
     app.get('/api/stats', asyncHandler(controller.getStats));
+    app.get('/api/version', controller.getVersionInfo);
     // Protected api
     const protected_api = passport.authenticate('bearer', { session: false });
     app.post('/api/auth', protected_api, okHandler);