+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 * Submitter email is used where PI email is expected ([#118](https://github.com/iomega/paired-data-form/issues/118))
+* Download button on pending page yields incorrectly formatted json files ([#120](https://github.com/iomega/paired-data-form/issues/120))
 
 ## [0.4.0] - 2020-03-20