+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Mention JSON schema on add form ([#115](https://github.com/iomega/paired-data-form/issues/115))
+* Project list can be sorted by clicking on column header ([#117](https://github.com/iomega/paired-data-form/issues/117))
+
+### Changed
+
+* Project list sorted on metabolite id ([#117](https://github.com/iomega/paired-data-form/issues/117))
 
 ### Fixed