+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0] - 2020-04-02
+
 ### Added
 
 * Mention JSON schema on add form ([#115](https://github.com/iomega/paired-data-form/issues/115))